Parse ssh URLs with port and different usernames
This can parse git@host:port/owner/repo and user@host:owner/repo name. I wanted to use ZnUrl but git is not using an absoluteURI. So

git@github.com:pharo-vcs/iceberg could not be parsed and I stay with 
the regexp. The notion of owner/project is not universally true either
but that is left for a future commit.