Use ZnUrl to parse HTTP URL
Use ZnUrl to parse the url. Add tests.  The notion of owner/project is
wrong as I can have git repos in many other places. E.g. in one of my
projects it is  http://git.osmocom.org/libosmo-sccp. Leave that for now.
path seems to be an empty ordered collection so do not write it.

Improves: #457