Add a Hiedra column to the Repository window.
I adds a delay when the window is opened because it is not asynchronous. But I think the information it adds makes it worth. 
In small branches it was a few milliseconds in my laptop, and the worst case I tested is 500ms slowdown with the dev-1.0 branch of the iceberg repository (>2800 commits).
The rendering part is not significant in performance because it uses a kind of pagination. For example, if a fasttable is opened and only the top 20 rows are visible, only one "page" of 30 rows are rendered and cached initially. When user scrolls the table, more pages of 30 rows are rendered on demand.
What could be optimized yet is the graph model: when the fasttable is opened, Hiedra currently iterates over all the commits and parents to create the model. But I don't plan to work in this optimization for the moment... so please test it.

IMPORTANT: It works in latest Pharo 8 + PR #2926 (https://github.com/pharo-project/pharo/pull/2926).