Initial and minimal support for OTP based 2factor authentication for github.
two (2) factor authentication is an extra security measure for a github account and requires providing a one time token as part of the authentication.  Add a minimal implementation (the bigger one requires oauth to create a long lived token). Add a subclass for the credentials that can store the tokens and then use it when approriate. Do double dispatch on it. Avoid slicing back to the default plain text type.

Details:
https://developer.github.com/v3/auth/#working-with-two-factor-authentication