Update tonel dependency to Pharo12 branch
During the development of Pharo 12 we will bring some changes to tonel such as an improved way to deal with packages and tags and maybe a way to set the deprecated aliases of a class. 

This change make Iceberg depend on the branch Pharo12 of the tonel repository. 

This branch has one change ahead of the master branch that is the introduction of TonelWriterV3. This new writer will export the package and the tag of classes on top of the category. 
It also update the reader to be able to read this new format and try to be as backward compatible as possible.