Improve diff performance by caching writerClass
The benefit of caching the writerClass is acknowledged in the comment of
IceDiff>>#writerClass, but that implementation is not effective. The
method #buildForChanges: will ultimately refer to the #writerClass of
the source and target, and do so repeatedly, per changed file in the
changes.

I made this change by profiling the following expression

```
myRepository branch diffTo: myRepository workingCopy
```

where the active branch of the repository is different from the
reference commit of the working copy. There are roughly a thousand
items in the collection passed to #buildForChanges:.

The duration drops from 9 to 6 seconds, more or less.